+++ CHANGES.md
@@ -33,6 +33,9 @@ master (in development)
   - Show chained checkers in Help buffers for syntax checkers [GH-571]
   - Map custom error levels to compilation mode levels [GH-700]
   - `flycheck-verify-checker` now includes the manually selected checker if any
+    [GH-705]
+  - `flycheck-select-checker` now shows a verification buffer if the selected
+    checker cannot be used [GH-705]
 
 - Bug fixes: